pyproject.toml
[project]
name = "pairs-trading"
version = "0.1.0"
description = ""
authors = [
    {name = "Angel Jara",email = "ajarac92@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.2.4,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "yfinance (>=0.2.55,<0.3.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alpaca-py (>=0.39.4,<0.40.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [{include = "pairs_trading", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"